+++ src/zoid/buttons/component.jsx
@@ -45,6 +45,8 @@ import {
   isPayPalTrustedUrl,
   getSDKInitTime,
   getSDKToken,
+  getShopperSessionId,
+  getGlobalSessionID,
 } from "@paypal/sdk-client/src";
 import {
   rememberFunding,
@@ -218,12 +220,19 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
       ).render(dom({ doc }));
     },
 
-    attributes: {
-      iframe: {
-        allowpaymentrequest: "allowpaymentrequest",
-        scrolling: "no",
-        title: FUNDING_BRAND_LABEL.PAYPAL,
-      },
+    attributes: ({ props }) => {
+      let fundingSource = "";
+      if (props.fundingSource) {
+        fundingSource = `-${props.fundingSource}`;
+      }
+
+      return {
+        iframe: {
+          allowpaymentrequest: "allowpaymentrequest",
+          scrolling: "no",
+          title: `${FUNDING_BRAND_LABEL.PAYPAL}${fundingSource}`,
+        },
+      };
     },
 
     eligible: ({ props }) => {
@@ -790,6 +799,12 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
         },
       },
 
+      globalSessionID: {
+        type: "string",
+        required: false,
+        value: getGlobalSessionID,
+      },
+
       hostedButtonId: {
         type: "string",
         required: false,
@@ -1178,6 +1193,12 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
         value: () => sessionState,
       },
 
+      shopperSessionId: {
+        type: "string",
+        required: false,
+        value: getShopperSessionId,
+      },
+
       getShopperInsightsUsed: {
         type: "function",
         value: () => wasShopperInsightsUsed,
@@ -1283,6 +1304,18 @@ export const getButtonsComponent: () => ButtonsComponent = memoize(() => {
         type: "object",
         required: false,
         default: () => window.__TEST_WALLET__,
+      },
+
+      hideSubmitButtonForCardForm: {
+        type: "boolean",
+        required: false,
+        queryParam: true,
+      },
+    },
+
+    exports: {
+      submitCardForm: {
+        type: "function",
       },
     },
   });